+++ app/models/abuse.rb
@@ -10,8 +10,20 @@ class Abuse < ApplicationRecord
   after_create :notify_admins_abuse_reported
 
   validates :first_name, :last_name, :email, :message, presence: true
+  validates :signaled_type, inclusion: { in: ['Project'], message: 'must be allowed type' }
+  validates :signaled_id, presence: true
+  validate :signaled_exists
 
   private
+
+  def signaled_exists
+    case signaled_type
+    when 'Project'
+      errors.add(:signaled_id, 'Project does not exist') unless Project.exists?(signaled_id)
+    else
+      errors.add(:signaled_type, 'Type does not allow')
+    end
+  end
 
   def notify_admins_abuse_reported
     NotificationCenter.call type: 'notify_admin_abuse_reported',